An attempt at a more consistent naming.
