+++ GPS_Distance.Tests/DistanceMeasurerTests/MeasureUsingHaversine_spec.cs
@@ -2,7 +2,7 @@
 using GPS_Distance.Models;
 using Xunit;
 using static GPS_Distance.Helpers.Helper;
-using static GPS_Distance.MeasurementFormulas.Measure;
+using static GPS_Distance.MeasurementFormulas.MeasureFormula;
 
 namespace DistanceMeasurerTests.HaversineFormula_spec
 {
@@ -25,7 +25,7 @@ namespace DistanceMeasurerTests.HaversineFormula_spec
             var endLocation = new Location(endLat, endLong);
 
             // Act
-            var actualDistance = HaversineFormula(startLocation, endLocation).ToUnit(Unit.Kilometres, 2);
+            var actualDistance = HaversineMeasure(startLocation, endLocation).ToUnit(Unit.Kilometres, 2);
 
             // Assert
             actualDistance.Should().Be(expectedDistance);
